first take on for purchase variations
